Add utilities file Move memory functions to utilities
