+++ system_metrics_collector/CMakeLists.txt
@@ -38,6 +38,7 @@ add_library(system_metrics_collector SHARED
   src/system_metrics_collector/linux_memory_measurement_node.cpp
   src/system_metrics_collector/periodic_measurement_node.cpp
   src/system_metrics_collector/proc_cpu_data.cpp
+  src/system_metrics_collector/utilities.cpp
   src/moving_average_statistics/moving_average.cpp
   src/moving_average_statistics/types.hpp
 )
