+++ system_metrics_collector/src/moving_average_statistics/moving_average.cpp
@@ -39,7 +39,7 @@ double MovingAverageStatistics::max() const
 
 double MovingAverageStatistics::min() const
 {
-  return MovingAverageStatistics::getStatistics().min;
+  return getStatistics().min;
 }
 
 double MovingAverageStatistics::standardDeviation() const
